Moves methods from routes to users controller
