+++ controllers/users_controller.js
@@ -6,7 +6,6 @@ const configuration = require('../knexfile')[environment];
 const database = require('knex')(configuration);
 var Food = require('../models/food')
 var Meal = require('../models/meal')
-// var User = require('../models/user')
 var pry = require('pryjs');
 var validator = require('email-validator');
 var session = require('../models/POJOs/session')
@@ -19,9 +18,6 @@ async function login(req,res) {
     if (validator.validate(email)) {
       const user = await database('users').where('email', email)
       if (user[0]) {
-        // let passwordHash = user.password
-        // let apiKey = user.apiKey
-        // let verify = bcrypt.compare(passwordAttempt, passwordHash)
         if (passwordAttempt === user[0].password) {
           session.setKey(user[0].apiKey)
           res.redirect('/users/welcome')
@@ -95,7 +91,73 @@ async function register(req,res) {
   }
 }
 
+async function foods() {
+  fetch('https://stormy-brushlands-92125.herokuapp.com/api/v1/foods')
+  .then(response => {
+    if (response.ok) {
+      return response.json();}
+      throw new Error('Request Failed.');},
+      networkError => console.log(networkError.message))
+      .then(foods => {
+        res.render('foods.ejs', {foods: foods})
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+}
+async function my_meals() {
+  fetch('https://stormy-brushlands-92125.herokuapp.com/api/v1/meals')
+  .then(response => {
+    if (response.ok) {
+      return response.json();}
+      throw new Error('Request Failed.');},
+      networkError => console.log(networkError.message))
+      .then(meals => {
+        let totalCalories = {}
+        meals.forEach(meal => {
+          totalCalories[meal.id] = 0
+          meal["foods"].forEach(food => {
+            totalCalories[meal.id] += parseInt(food["calories"])
+          })
+        })
+        res.render('my_meals.ejs', {
+          meal: undefined, //this is bad
+          meals: meals,
+          totalCalories: totalCalories
+        })
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+}
+async function my_meal () {
+  fetch(`https://stormy-brushlands-92125.herokuapp.com/api/v1/meals/${req.params.id}/foods`)
+  .then(response => {
+    if (response.ok) {
+      return response.json();}
+      throw new Error('Request Failed.');},
+      networkError => console.log(networkError.message))
+      .then(meal => {
+        let totalCalories = 0;
+        meal["foods"].forEach(food => {
+          totalCalories += parseInt(food["calories"])
+        })
+        res.render('my_meals.ejs', {
+          meals: undefined, //this is bad
+          meal: meal,
+          totalCalories: totalCalories,
+          foods: meal["foods"]
+        })
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+}
+
 module.exports = {
+  foods: foods,
+  my_meals: my_meals,
+  my_meal: my_meal,
   login: login,
   register: register
 }
